renamed package, update example
