+++ example/cavity/precice-config.xml
@@ -67,7 +67,7 @@
     <coupling-scheme:serial-implicit>
 
         <time-window-size value="0.05"/>
-        <max-time value="70"/>
+        <max-time value="30"/>
 
         <participants first="Fluid_Solver" second="Structure_Solver"/>
 
@@ -85,23 +85,24 @@
         <max-iterations value="100"/>
 
         <relative-convergence-measure
-            limit="1.0e-4"
+            limit="1.0e-3"
             data="DisplacementDelta"
             mesh="Structure_Nodes"/>
         <relative-convergence-measure
-            limit="1.0e-4"
+            limit="1.0e-3"
             data="Force"
             mesh="Structure_Nodes"/>
 
         <!-- <extrapolation-order value="2"/> -->
 
-        <acceleration:IQN-IMVJ>
+        <acceleration:IQN-ILS>
             <data name="DisplacementDelta" mesh="Structure_Nodes"/>
             <preconditioner type="residual-sum"/>
+            <filter type="QR2" limit="1.2e-3"/>
             <initial-relaxation value="0.1"/>
             <max-used-iterations value="60"/>
             <time-windows-reused value="15"/>
-        </acceleration:IQN-IMVJ>
+        </acceleration:IQN-ILS>
 
     </coupling-scheme:serial-implicit>
