+++ setup.cfg
@@ -1,5 +1,5 @@
 [metadata]
-name = mbdyn-adapter
+name = mbdynprecice
 version = 0.0.2
 author = Mikko Folkersma, Hagen Leipprand
 author_email = hagen.leipprand@gmail.com
@@ -16,7 +16,6 @@ classifiers =
     Intended Audience :: Science/Research
     License :: OSI Approved :: GNU General Public License v3 (GPLv3)
     Topic :: Scientific/Engineering
-
 
 [options]
 package_dir =
